Factor handle into treasury_email_common.py
